BoundaryConditions

The BoundaryConditions module is used to initialize the integrated boundary condition model for the Stokes-continuity and heat equations. An integrated boundary condition model also manages volume conservation and marker recycling for inflow/outflow boundary conditions.

Boundary Condition Models

LithosphericExtensionMovingBoundaries

LithosphericExtensionFixedBoundaries

LithosphericExtensionFixedBoundariesAsymmetric

LithosphericExtensionDepthDependent

LithosphericExtensionInflowAndOutflowAlongSides

ExtensionNoBottomInflow

PureShearDeformation

VerticalCouetteFlow

SolidBodyRotation

RayleighTaylorInstability

BoxConvection

ElasticSlab

ViscousBlock

SandboxShortening

SandboxExtension

VerticalChannelFlowNonNewtonian

VerticalChannelFlowIsoviscous

VerticalChannelFlowShearHeating

FractureZoneContractionFixedBoundaries

FractureZoneContractionFixedBoundariesAsymmetric

ViscoelasticContraction

ViscoelasticContractionAsymmetric

quick search for boundary condition model type

Highlight a boundary condition model name in the list above and use Ctl-F + Enter to navigate to a detailed description in the Integrated Boundary Condition Model Types section below.

Initialization

EarthBox.BoundaryConditions.initialize! — Function

initialize!(
    model::ModelData;
    model_type::Union{Int, String, Symbol, Nothing}=nothing
)::Nothing

Initialize integrated boundary condition model type for coupled Stokes-continuity and heat equations.

Arguments

model::ModelData: The model data object containing model parameters and arrays.

model_type::Union{Int, String, Symbol, Nothing}: Controls the type of integrated boundary condition model type for the Stokes-continuity and heat equations. See the Integrated Boundary Condition Model Types section below for information on available boundary condition model types. The boundary condition model type is stored in the model data container as an integer ID (itype_bc) and a corresponding string name (stype_bc). If model_type is nothing the current boundary condition type defined in the model data container will be used. The boundary condition model type parameters can be accessed from the model data container as follows:

itype_bc = model.bcs.parameters.bc_options.itype_bc.value

stype_bc = model.bcs.parameters.bc_options.stype_bc.value

Returns

Nothing

Integrated Boundary Condition Model Types

LithosphericExtensionMovingBoundaries

model_type value: "LithosphericExtensionMovingBoundaries", :LithosphericExtensionMovingBoundaries, or 0

Description: This boundary condition type is used for extension models with moving boundaries. A fixed free slip upper boundary is used with moving lower and side boundaries. Orthogonal velocity is prescribed along side and bottom boundaries. Regridding is applied with no marker recycling since boundaries are moving. Temperature is prescribed on the top and bottom, and zero heat flux is applied along side boundaries.

LithosphericExtensionFixedBoundaries

model_type value: "LithosphericExtensionFixedBoundaries", :LithosphericExtensionFixedBoundaries, or 1

Description: This boundary condition type is for extension models with fixed boundaries. Free slip with orthogonal velocity is prescribed along top, bottom and side boundaries to ensure volume conservation. Markers are recycled with special treatment in the sticky air region to ensure constant sticky air volume. Temperature is prescribed on the top and bottom, and zero heat flux is applied along side boundaries.

LithosphericExtensionFixedBoundariesAsymmetric

model_type value: "LithosphericExtensionFixedBoundariesAsymmetric", :LithosphericExtensionFixedBoundariesAsymmetric, or 2

Description: This boundary condition type is for extension models with fixed boundaries and asymmetric boundary conditions. Free slip in the y-direction and zero lateral orthogonal velocity are applied on the left boundary, and uniform lateral full extension velocity with free slip in y-direction is applied along the right boundary. Free slip with orthogonal velocity is prescribed along top and bottom boundaries to ensure volume conservation. Markers are recycled with special treatment in the sticky air region to ensure constant sticky air volume. Temperature is prescribed on the top and bottom, and zero heat flux is applied along side boundaries.

LithosphericExtensionDepthDependent

model_type value: "LithosphericExtensionDepthDependent", :LithosphericExtensionDepthDependent, or 3

Description: This boundary condition type is used for extension models with fixed boundaries and side boundaries with free slip and depth-dependent prescribed orthogonal velocity. Along top and bottom boundaries free slip is applied with prescribed orthogonal velocity calculated to ensure volume conservation. Particles are recycled (no regridding) with special treatment in sticky air region to ensure constant sticky air volume. Temp. boundary conditions on top and bottom, zero flux on sides.

LithosphericExtensionInflowAndOutflowAlongSides

model_type value: "LithosphericExtensionInflowAndOutflowAlongSides", :LithosphericExtensionInflowAndOutflowAlongSides, or 4

Description: This boundary condition type is used for extension models with fixed boundaries and side boundaries with free slip and depth-dependent prescribed orthogonal velocity for inflow and outflow. Along top and bottom boundary free slip is applied with prescribed orthogonal velocity calculated to ensure volume conservation in the sticky layer. Particles are recycled (no regridding) with special treatment in sticky and side inflow regions to ensure. Temperature boundary conditions on top and bottom, zero flux on sides.

ExtensionNoBottomInflow

model_type value: "ExtensionNoBottomInflow", :ExtensionNoBottomInflow, or 5

Description: This boundary condition type is used for the simple extension with constant outflow along the side boundary, conservative inflow along the top boundary and a closed free slip boundary along the bottom. Velocity is prescribed along side and top boundaries to ensure the conservation of volume.

PureShearDeformation

model_type value: "PureShearDeformation", :PureShearDeformation, or 6

Description: This boundary conditions type is used for a pure shear experiment. Free slip is applied along boundaries with orthogonal prescribed velocity. Temperature is prescribed on top, bottom and side boundaries. No marker recycling is used.

VerticalCouetteFlow

model_type value: "VerticalCouetteFlow", :VerticalCouetteFlow, or 7

Description: Boundary conditions for vertical Couette flow (simple shear). Open channel boundary conditions are applied along the top and bottom boundaries to model vertical flow in an infinite channel. This is accomplished by prescribing zero vertical velocity gradients along top and bottom boundaries and by prescribing a vertical pressure gradient, which for this case is zero. No-slip bcs on are applied on side boundaries with prescribed y-velocity along the right boundary to induce shear flow. Zero vertical temperature gradients are prescribed along top and bottom boundaries. Prescribed temperature is applied along left boundary. An insulating boundary condition (zero lateral heat flux) is implemented along the right boundary. Marker recycling takes into account vertical channel flow.

SolidBodyRotation

model_type value: "SolidBodyRotation", :SolidBodyRotation, or 8

Description: This boundary condition type is used for the solid body rotation benchmark with fixed boundaries (Stokes equations are not solved). Prescribed thermal gradients are applied along side boundaries.

RayleighTaylorInstability

model_type value: "RayleighTaylorInstability", :RayleighTaylorInstability, or 9

Description: This boundary condition type is used for the Rayleigh-Taylor instability benchmark from Ramberg (1968). No slip boundary conditions are applied on top and bottom boundaries and free slip boundary condition are applied on side boundaries.

BoxConvection

model_type value: "BoxConvection", :BoxConvection, or 10

Description: This boundary conditions type is used for the convection in a box benchmark. Free slip boundary conditions are applied on all sides. Temperature is prescribed on top and bottom boundaries. Side boundaries have insulating thermal boundary conditions.

ElasticSlab

model_type value: "ElasticSlab", :ElasticSlab, or 11

Description: This boundary conditions type is used for the elastic slab experiment. No slip boundary conditions are applied on the left boundary. Free slip boundary conditions are applied on top, bottom and right boundaries.

ViscousBlock

model_type value: "ViscousBlock", :ViscousBlock, or 12

Description: This boundary conditions type is used for the viscous block benchmark. Free slip boundary conditions are applied along all boundaries.

SandboxShortening

model_type value: "SandboxShortening", :SandboxShortening, or 13

Description: This boundary condition type is used for the sandbox shortening benchmark. No slip boundary conditions are applied on bottom and left boundaries. Free slip boundary conditions are applied on top and right boundaries. Prescribed internal velocity boundary conditions are used to model a mobile wall. Boundary friction is also used along the left and bottom boundaries in addition to the interface between sand and mobile wall. Plastic failure is prevented from occurring in the mobile wall.

SandboxExtension

model_type value: "SandboxExtension", :SandboxExtension, or 14

Description: This boundary condition type is used for the sandbox extension benchmark. No slip boundary conditions are applied on bottom and left boundaries. Free slip boundary conditions are applied on top and right boundaries. Prescribed internal velocity boundary conditions are used to model a mobile wall. Boundary friction is also used along the left and bottom boundaries in addition to the interface between sand and mobile wall. Plastic failure is prevented from occurring in the mobile wall and thin plate extending from mobile wall.

VerticalChannelFlowNonNewtonian

model_type value: "VerticalChannelFlowNonNewtonian", :VerticalChannelFlowNonNewtonian, or 16

Description: This boundary condition type is used for vertical channel flow with prescribed pressure and velocity gradients along upper and lower boundaries. No-slip boundary condition are applied on side boundaries. Prescribed temperature is applied along top and bottom boundaries. Prescribed gradients are applied along side boundaries. Special regridding for is applied for channel flow. This bc type is used for non-newtonian channel flow benchmark with temperature along side boundaries set equal to zero.

VerticalChannelFlowIsoviscous

model_type value: "VerticalChannelFlowIsoviscous", :VerticalChannelFlowIsoviscous, or 15

Description: This boundary conditions type is used for vertical channel flow with prescribed pressure and velocity gradients along upper and lower boundaries. No-slip boundary conditions are applied on side boundaries. Prescribed temperature along top and bottom boundaries. Prescribed gradients are applied along side boundaries. Special regridding for is applied for channel flow. This bc type is used for the isoviscous channel flow benchmark with thermal advection benchmark and temperature prescribed along side boundaries.

VerticalChannelFlowShearHeating

model_type value: "VerticalChannelFlowShearHeating", :VerticalChannelFlowShearHeating, or 17

Description: This boundary conditions type is used for vertical channel flow with prescribed pressure and velocity gradients along upper and lower boundaries. No-slip boundary conditions are applied on side boundaries. Prescribed temperature along top and bottom boundaries. Prescribed gradients are applied along side boundaries. Special regridding for is applied for channel flow. This boundary condition type is used for the channel flow benchmark with variable thermal conductivity and shear heating.

FractureZoneContractionFixedBoundaries

model_type value: "FractureZoneContractionFixedBoundaries", :FractureZoneContractionFixedBoundaries, or 18

Description: This boundary condition type is used for contraction models with fixed boundaries, symmetric contraction and fracture zone material setup. Free slip boundary conditions are applied along top, bottom and side boundaries. Orthogonal velocity is prescribed along side boundaries. Temperature is prescribed on top and bottom, and zero heat flux is applied along side boundaries.

FractureZoneContractionFixedBoundariesAsymmetric

model_type value: "FractureZoneContractionFixedBoundariesAsymmetric", :FractureZoneContractionFixedBoundariesAsymmetric, or 19

Description: This boundary condition type is used for contraction models with fixed boundaries, asymmetric contraction and fracture zone material setup. Free slip boundary conditions are applied along top, bottom and side boundaries. Orthogonal velocity is prescribed along side boundaries with zero inflow along the left boundary. Temperature is prescribed on top and bottom, and zero heat flux is applied along side boundaries.

ViscoelasticContraction

model_type value: "ViscoelasticContraction", :ViscoelasticContraction, or 20

Description: This boundary condition type is used for contraction models with fixed boundaries, symmetric contraction and viscoelastic lithosphere material setup. Free slip boundary conditions are applied along top, bottom and side boundaries. Orthogonal velocity is prescribed along side boundaries. Temperature is prescribed on top and bottom and zero heat flux is applied along side boundaries.

ViscoelasticContractionAsymmetric

model_type value: "ViscoelasticContractionAsymmetric", :ViscoelasticContractionAsymmetric, or 21

Description: This boundary condition type is used for contraction models with fixed boundaries, asymmetric contraction and viscoelastic lithosphere material setup. Free slip boundary conditions are applied along top, bottom and side boundaries. Orthogonal velocity is prescribed along side boundaries with zero inflow along the left boundary. Temperature is prescribed on top and bottom and zero heat flux is applied along side boundaries.

Example