Material Properties

The following docstring contains a list of material properties that can be defined for each material in material library files. See Material Collection Files for more details.

EarthBox.Markers.MarkerMaterials.MaterialsContainer.MaterialProperties.MaterialPropertiesState — Type

MaterialPropertiesState

Data structure for storing material properties state.

Fields

melting_params::MeltingParameters: Melting parameters

material_color::MaterialColor: Material color

density::Density: Density

dislocation_creep::DislocationCreep: Dislocation creep

diffusion_creep::DiffusionCreep: Diffusion creep

peierls_creep::PeierlsCreep: Peierls creep

temperature_dependent_viscosity::TemperatureDependentViscosity: Temperature dependent viscosity

blankenbach89_viscosity::Blankenbach89Viscosity: Blankenbach89 viscosity

shear_modulus::ShearModulus: Shear modulus

plasticity::Plasticity: Plasticity

flow_law::FlowLaw: Flow law

heat_capacity::HeatCapacity: Heat capacity

thermal_conductivity::ThermalConductivity: Thermal conductivity

radiogenic_heat_production::RadiogenicHeatProduction: Radiogenic heat production

compaction::Compaction: Compaction

Returns

MaterialPropertiesState: New MaterialPropertiesState parameter group with initialized values

List of Material Properties For Material Collection Files

For a given material in a material collection library file enter the name of the properties followed by a list containing value, units and description:

MaterialName:
    property_name_1: [value, units, description]
    property_name_2: [value, units, description]

Use the standard units defined below. A future version of EarthBox will support material property unit conversion.

The following properties can be entered for a given material:

MeltingParameters

itype_liquidus: Liquidus type ID

stype_liquidus: Liquidus type name

itype_solidus: Solidus type ID

stype_solidus: Solidus type name

latent_heat: Latent heat in J/kg

Density

standard_density: Standard density in kg/m^3

thermal_expansion: Thermal expansion coefficient in 1/K

compressibility: Compressibility in 1/Pa

melt_density: Melt density in kg/m^3

DislocationCreep

pre_exponential_dc: Pre-exponential factor for dislocation creep in 1/s/MPa^n

stress_exponent_n_dc: Stress exponent for dislocation creep

activation_energy_dc: Activation energy for dislocation creep in kJ/mol

activation_volume_dc: Activation volume for dislocation creep in J/MPa/mol

DiffusionCreep

pre_exponential_difc: Pre-exponential factor for diffusion creep in 1/s/MPa^n

activation_energy_difc: Activation energy for diffusion creep in kJ/mol

activation_volume_difc: Activation volume for diffusion creep in J/MPa/mol

PeierlsCreep

pre_exponential_pei: Pre-exponential factor for Peierls creep in s^-m1*MPa^-m2

stress_exponent_m1_pei: Stress exponent m1 for Peierls creep

stress_exponent_m2_pei: Stress exponent m2 for Peierls creep

peierls_stress: Peierls stress in MPa

TemperatureDependentViscosity

pre_exponential_td: Pre-exponential factor for temperature dependent viscosity in Pa.s

activation_energy_td: Activation energy for temperature dependent viscosity in kJ/mol

Blankenbach89Viscosity

viscosity_ref_blankenbach89: Reference viscosity for Blankenbach89 in Pa.s

b_term_blankenbach89: b term for Blankenbach89

c_term_blankenbach89: c term for Blankenbach89

ShearModulus

shear_modulus: Shear modulus in Pa

Plasticity

cohesion_initial: Initial cohesion in Pa

cohesion_final: Final cohesion in Pa

friction_angle_initial: Initial friction angle in degrees

friction_angle_final: Final friction angle in degrees

strain_initial: Initial strain

strain_final: Final strain

dilatation_angle: Dilatation angle in degrees

FlowLaw

flow_type: Flow law type

flow_stype: Flow law type name

HeatCapacity

heat_capacity: Heat capacity in J/kg/K

ThermalConductivity

thermal_conductivity_ref: Reference thermal conductivity in W/m/K

thermal_conductivity_a: Thermal conductivity parameter a in W/m

RadiogenicHeatProduction

radiogenic_heat_production: Radiogenic heat production in W/m^3

Compaction

porosity_initial: Initial porosity in fraction

porosity_decay_depth: Porosity decay depth in m

MeltingParameters

EarthBox.Markers.MarkerMaterials.MaterialsContainer.MaterialProperties.MeltingParameters — Type

MeltingParameters

Data structure for storing melting parameters.

Fields

itype_liquidus::ParameterInt: Liquidus type ID

stype_liquidus::ParameterStr: Liquidus type name

itype_solidus::ParameterInt: Solidus type ID

stype_solidus::ParameterStr: Solidus type name

latent_heat::ParameterFloat: Latent heat in J/kg

Returns

MeltingParameters: New MeltingParameters parameter group with initialized values

Solidus Models

stype_solidus	itype_solidus	Description

"NoMelting"	-1	Negative one is used to signal a material without melting. This is the default value for itype_solidus.

"SedimentGerya2010"	0	Sediment solidus model from Gerya (2010) Table 17.2.

"UprContCrustGerya2010"	1	Upper continental crust solidus model from Gerya (2010) Table 17.2.

"LwrContCrustGerya2010"	2	Lower continental crust solidus model from Gerya (2010) Table 17.2.

"DryMantleGerya2010"	3	Dry mantle solidus model from Gerya (2010) Table 17.2.

"PeridotiteKatz2003"	4	Peridotite solidus model from Katz et al. (2003).

"GabbroGerya2010"	5	Gabbro solidus model from Gerya (2010) Table 17.2.

"GabbroSchubert2013"	6	Gabbro solidus model from Schubert et al. (2013)

"GabbroGlacier"	7	Gabbro solidus model that approximates the gabbro glacier model of Maclennan et al. (2004).

Liquidus Models

stype_liquidus	itype_liquidus	Description

"NoMelting"	-1	Negative one is used to signal a material without melting. This is the default value for itype_liquidus.

"SedimentGerya2010"	0	Sediment liquidus model from Gerya (2010) Table 17.2.

"UprContCrustGerya2010"	1	Upper continental crust liquidus model from Gerya (2010) Table 17.2.

"LwrContCrustGerya2010"	2	Lower continental crust liquidus model from Gerya (2010) Table 17.2.

"DryMantleGerya2010"	3	Dry mantle liquidus model from Gerya (2010) Table 17.2.

"PeridotiteKatz2003"	4	Peridotite liquidus model from Katz et al. (2003).

"GabbroGerya2010"	5	Gabbro liquidus model from Gerya (2010) Table 17.2.

"GabbroGlacier"	6	Gabbro liquidus model that approximates the gabbro glacier model of Maclennan et al. (2004).

MaterialColor

EarthBox.Markers.MarkerMaterials.MaterialsContainer.MaterialProperties.MaterialColor — Type

MaterialColor

Data structure for storing material color parameters.

Fields

red_fraction::ParameterFloat: Red color fraction

green_fraction::ParameterFloat: Green color fraction

blue_fraction::ParameterFloat: Blue color fraction

Returns

MaterialColor: New MaterialColor parameter group with initialized values

Density

EarthBox.Markers.MarkerMaterials.MaterialsContainer.MaterialProperties.Density — Type

Density

Data structure for storing density parameters.

Fields

standard_density::ParameterFloat: Standard density in kg/m^3

thermal_expansion::ParameterFloat: Thermal expansion coefficient in 1/K

compressibility::ParameterFloat: Compressibility in 1/Pa

melt_density::ParameterFloat: Melt density in kg/m^3

Returns

Density: New Density parameter group with initialized values

DislocationCreep

EarthBox.Markers.MarkerMaterials.MaterialsContainer.MaterialProperties.DislocationCreep — Type

DislocationCreep

Data structure for storing dislocation creep parameters.

Fields

pre_exponential_dc::ParameterFloat: Pre-exponential factor for dislocation creep in 1/s/MPa^n

stress_exponent_n_dc::ParameterFloat: Stress exponent for dislocation creep

activation_energy_dc::ParameterFloat: Activation energy for dislocation creep in kJ/mol

activation_volume_dc::ParameterFloat: Activation volume for dislocation creep in J/MPa/mol

Returns

DislocationCreep: New DislocationCreep parameter group with initialized values

DiffusionCreep

EarthBox.Markers.MarkerMaterials.MaterialsContainer.MaterialProperties.DiffusionCreep — Type

DiffusionCreep

Data structure for storing diffusion creep parameters.

Fields

pre_exponential_difc::ParameterFloat: Pre-exponential factor for diffusion creep in 1/s/MPa^n

activation_energy_difc::ParameterFloat: Activation energy for diffusion creep in kJ/mol

activation_volume_difc::ParameterFloat: Activation volume for diffusion creep in J/MPa/mol

Returns

DiffusionCreep: New DiffusionCreep parameter group with initialized values

PeierlsCreep

EarthBox.Markers.MarkerMaterials.MaterialsContainer.MaterialProperties.PeierlsCreep — Type

PeierlsCreep

Data structure for storing peierls creep parameters.

Fields

pre_exponential_pei::ParameterFloat: Pre-exponential factor for Peierls creep in s^-m1*MPa^-m2

stress_exponent_m1_pei::ParameterFloat: Stress exponent m1 for Peierls creep

stress_exponent_m2_pei::ParameterFloat: Stress exponent m2 for Peierls creep

peierls_stress::ParameterFloat: Peierls stress in MPa

Returns

PeierlsCreep: New PeierlsCreep parameter group with initialized values

TemperatureDependentViscosity

EarthBox.Markers.MarkerMaterials.MaterialsContainer.MaterialProperties.TemperatureDependentViscosity — Type

TemperatureDependentViscosity

Data structure for storing temperature dependent viscosity parameters.

Fields

pre_exponential_td::ParameterFloat: Pre-exponential factor for temperature dependent viscosity in Pa.s

activation_energy_td::ParameterFloat: Activation energy for temperature dependent viscosity in kJ/mol

Returns

TemperatureDependentViscosity: New TemperatureDependentViscosity parameter group with initialized values

Blankenbach89Viscosity

EarthBox.Markers.MarkerMaterials.MaterialsContainer.MaterialProperties.Blankenbach89Viscosity — Type

Blankenbach89Viscosity

Data structure for storing blankenbach89 viscosity parameters.

Fields

viscosity_ref_blankenbach89::ParameterFloat: Reference viscosity for Blankenbach89 in Pa.s

b_term_blankenbach89::ParameterFloat: b term for Blankenbach89

c_term_blankenbach89::ParameterFloat: c term for Blankenbach89

Returns

Blankenbach89Viscosity: New Blankenbach89Viscosity parameter group with initialized values

ShearModulus

EarthBox.Markers.MarkerMaterials.MaterialsContainer.MaterialProperties.ShearModulus — Type

ShearModulus

Data structure for storing shear modulus parameters.

Fields

shear_modulus::ParameterFloat: Shear modulus in Pa

Returns

ShearModulus: New ShearModulus parameter group with initialized values

Plasticity

EarthBox.Markers.MarkerMaterials.MaterialsContainer.MaterialProperties.Plasticity — Type

Plasticity

Data structure for storing plasticity parameters.

Fields

cohesion_initial::ParameterFloat: Initial cohesion in Pa

cohesion_final::ParameterFloat: Final cohesion in Pa

friction_angle_initial::ParameterFloat: Initial friction angle in degrees

friction_angle_final::ParameterFloat: Final friction angle in degrees

sine_friction_angle_initial::ParameterFloat: Initial sine friction angle

sine_friction_angle_final::ParameterFloat: Final sine friction angle

strain_initial::ParameterFloat: Initial strain

strain_final::ParameterFloat: Final strain

dilatation_angle::ParameterFloat: Dilatation angle in degrees

Returns

Plasticity: New Plasticity parameter group with initialized values

FlowLaw

EarthBox.Markers.MarkerMaterials.MaterialsContainer.MaterialProperties.FlowLaw — Type

FlowLaw

Data structure for storing flow law parameters.

Fields

flow_type::ParameterInt: Flow law type

flow_stype::ParameterStr: Flow law type name

viscosity_iso::ParameterFloat: Isoviscous viscosity in Pa.s

dislocation_creep::DislocationCreep: DislocationCreep parameter group

peierls_creep::PeierlsCreep: PeierlsCreep parameter group

diffusion_creep::DiffusionCreep: DiffusionCreep parameter group

temperature_dependent_viscosity::TemperatureDependentViscosity: TemperatureDependentViscosity parameter group

blankenbach89_viscosity::Blankenbach89Viscosity: Blankenbach89Viscosity parameter group

plasticity::Plasticity: Plasticity parameter group

shear_modulus::ShearModulus: ShearModulus parameter group

Returns

FlowLaw: New FlowLaw parameter group with initialized values

HeatCapacity

EarthBox.Markers.MarkerMaterials.MaterialsContainer.MaterialProperties.HeatCapacity — Type

HeatCapacity

Data structure for storing heat capacity parameters.

Fields

heat_capacity::ParameterFloat: Heat capacity in J/kg/K

Returns

HeatCapacity: New HeatCapacity parameter group with initialized values

ThermalConductivity

EarthBox.Markers.MarkerMaterials.MaterialsContainer.MaterialProperties.ThermalConductivity — Type

ThermalConductivity

Data structure for storing thermal conductivity parameters.

Fields

thermal_conductivity_ref::ParameterFloat: Reference thermal conductivity in W/m/K

thermal_conductivity_a::ParameterFloat: Thermal conductivity parameter a in W/m

Returns

ThermalConductivity: New ThermalConductivity parameter group with initialized values

RadiogenicHeatProduction

EarthBox.Markers.MarkerMaterials.MaterialsContainer.MaterialProperties.RadiogenicHeatProduction — Type

RadiogenicHeatProduction

Data structure for storing radiogenic heat production parameters.

Fields

radiogenic_heat_production::ParameterFloat: Radiogenic heat production in W/m^3

Returns

RadiogenicHeatProduction: New RadiogenicHeatProduction parameter group with initialized values

Compaction

EarthBox.Markers.MarkerMaterials.MaterialsContainer.MaterialProperties.Compaction — Type

Compaction

Data structure for storing compaction parameters.

Fields

porosity_initial::ParameterFloat: Initial porosity in fraction

porosity_decay_depth::ParameterFloat: Porosity decay depth in m

Returns

Compaction: New Compaction parameter group with initialized values

Solidus Models

Liquidus Models