FractureZone

The FractureZone material geometry is composed of lithospheric domains with different thicknesses separated by a weak zone superimposed by oceanic crust and a layer of sediment.

Initialization

EarthBox.MaterialGeometry.FractureZone.initialize! — Function

Initialize fracture zone geometry.

Keyword arguments:

sediment_thickness::Float64:

Thickness in meters of sediment layer

basaltic_oceanic_crust_thickness::Float64:

Thickness in meters of basaltic oceanic crust layer

gabbroic_oceanic_crust_thickness::Float64:

Thickness in meters of gabbroic oceanic crust layer

thickness_of_younger_lithosphere::Float64:

Thickness in meters of younger lithosphere

thickness_of_older_lithosphere::Float64:

Thickness in meters of older lithosphere

thickness_of_weak_lithosphere::Float64:

Thickness in meters of weak lithosphere

x_fracture_zone_start::Float64:

Start of fracture zone in x-direction in meters

x_fracture_zone_end::Float64:

End of fracture zone in x-direction in meters