Geometry

Parameters

StickyAirGeometry

EarthLayering

WeakSeed

WeakFault

LithoStrongZones

Plume

CrustalHole

RayleighTaylor

MobileWall

Sandbox

InternalVelocityZone

FractureZone

EarthBox.ModelDataContainer.GeometryContainer.Geometry — Type

Geometry <: CollectionContainer

Struct containing material geometry parameters.

Fields

parameters::Parameters: Parameter groups for material geometry.

EarthBox.ModelDataContainer.GeometryContainer.ParameterCollection.Parameters — Type

Parameters <: AbstractParameterCollection

Collection of geometry parameters.

Fields

sticky_air_geometry::StickyAirGeometry: Sticky air layer geometry parameters

earth_layering::EarthLayering: Earth layering parameters

weak_seed::WeakSeed: Weak seed parameters

weak_fault::WeakFault: Weak fault parameters

litho_strong_zones::LithoStrongZones: Lithospheric strong zone parameters

plume::Plume: Mantle plume parameters

crustal_hole::CrustalHole: Crustal hole parameters

rayleigh_taylor::RayleighTaylor: Rayleigh-Taylor instability parameters

mobile_wall::MobileWall: Mobile wall parameters

sandbox::Sandbox: Sandbox experiment parameters

internal_velocity_zone::InternalVelocityZone: Internal velocity zone parameters

fracture_zone::FractureZone: Fracture zone parameters

EarthBox.ModelDataContainer.GeometryContainer.ParameterCollection.StickyAirGeometryGroup.StickyAirGeometry — Type

StickyAirGeometry <: AbstractParameterGroup

Parameter group for sticky air geometry parameters.

Fields

thick_air::ParameterFloat: Thickness in meters of sticky air in meters

obj_list::Vector{ParameterFloat}: List of parameter objects

Nested Dot Access

thick_air = model.geometry.parameters.sticky_air_geometry.thick_air.value

Constructor

StickyAirGeometry()

Create a new StickyAirGeometry parameter group with default values.

Returns

StickyAirGeometry: New StickyAirGeometry parameter group with initialized default values

EarthBox.ModelDataContainer.GeometryContainer.ParameterCollection.EarthLayeringGroup.EarthLayering — Type

EarthLayering <: AbstractParameterGroup

Parameter group for Earth layering geometry parameters.

Fields

thick_upper_crust::ParameterFloat: Thickness in meters of upper crust

thick_lower_crust::ParameterFloat: Thickness in meters of lower crust

thick_upper_lith::ParameterFloat: Thickness in meters of upper mantle lithosphere

thick_middle_lith::ParameterFloat: Thickness in meters of middle lithosphere

thick_lower_lith::ParameterFloat: Thickness in meters of lower mantle lithosphere

thick_lith::ParameterFloat: Thickness in meters of lithosphere

thick_crust::ParameterFloat: Thickness in meters of crust

thick_mantle_lith::ParameterFloat: Thickness in meters of mantle lithosphere

obj_list::Vector{Union{ParameterFloat, ParameterInt}}: List of parameter objects

Nested Dot Access

thick_upper_crust = model.geometry.parameters.earth_layering.thick_upper_crust.value

thick_lower_crust = model.geometry.parameters.earth_layering.thick_lower_crust.value

thick_upper_lith = model.geometry.parameters.earth_layering.thick_upper_lith.value

thick_middle_lith = model.geometry.parameters.earth_layering.thick_middle_lith.value

thick_lower_lith = model.geometry.parameters.earth_layering.thick_lower_lith.value

thick_lith = model.geometry.parameters.earth_layering.thick_lith.value

thick_crust = model.geometry.parameters.earth_layering.thick_crust.value

thick_mantle_lith = model.geometry.parameters.earth_layering.thick_mantle_lith.value

Constructor

EarthLayering()

Create a new EarthLayering parameter group with default values (NaN indicates unset parameters).

Returns

EarthLayering: New EarthLayering parameter group with initialized default values

EarthBox.ModelDataContainer.GeometryContainer.ParameterCollection.WeakSeedGroup.WeakSeed — Type

WeakSeed <: AbstractParameterGroup

Parameter group for weak seed geometry parameters.

Fields

w_seed::ParameterFloat: Width in meters of weak seed

x_seed::ParameterFloat: x-location in meters of weak seed

y_seed::ParameterFloat: y-location in meters of weak seed

iuse_weak_seed::ParameterInt: Use weak seed (0=off, 1=on)

obj_list::Vector{Union{ParameterFloat, ParameterInt}}: List of parameter objects

Nested Dot Access

w_seed = model.geometry.parameters.weak_seed.w_seed.value

x_seed = model.geometry.parameters.weak_seed.x_seed.value

y_seed = model.geometry.parameters.weak_seed.y_seed.value

iuse_weak_seed = model.geometry.parameters.weak_seed.iuse_weak_seed.value

Constructor

WeakSeed()

Create a new WeakSeed parameter group with default values.

Returns

WeakSeed: New WeakSeed parameter group with initialized default values

EarthBox.ModelDataContainer.GeometryContainer.ParameterCollection.WeakFaultGroup.WeakFault — Type

WeakFault <: AbstractParameterGroup

Parameter group for weak fault geometry parameters.

Fields

fault_dip_degrees::ParameterFloat: Dip in degrees of the weak zone used to approximate a fault

fault_thickness::ParameterFloat: Thickness in meters of the weak zone used to approximate a fault

x_initial_fault::ParameterFloat: Initial x-coordinate in meters of the weak zone used to approximate a fault

fault_height::ParameterFloat: Height in meters of the weak zone used to approximate a fault

iuse_weak_fault::ParameterInt: Use weak fault (0=off, 1=on)

obj_list::Vector{Union{ParameterFloat, ParameterInt}}: List of parameter objects

Nested Dot Access

fault_dip_degrees = model.geometry.parameters.weak_fault.fault_dip_degrees.value

fault_thickness = model.geometry.parameters.weak_fault.fault_thickness.value

x_initial_fault = model.geometry.parameters.weak_fault.x_initial_fault.value

fault_height = model.geometry.parameters.weak_fault.fault_height.value

iuse_weak_fault = model.geometry.parameters.weak_fault.iuse_weak_fault.value

Constructor

WeakFault()

Create a new WeakFault parameter group with default values.

Returns

WeakFault: New WeakFault parameter group with initialized default values

EarthBox.ModelDataContainer.GeometryContainer.ParameterCollection.LithoStrongZonesGroup.LithoStrongZones — Type

LithoStrongZones <: AbstractParameterGroup

Parameter group for lithospheric strong zones geometry parameters.

Fields

x_left_strong::ParameterFloat: x-coordinate in meters of the left edge of the lithospheric strong zone

x_right_strong::ParameterFloat: x-coordinate in meters of the right edge of the lithospheric strong zone

iuse_strong_zones::ParameterInt: Use strong zones (0=off, 1=on)

obj_list::Vector{Union{ParameterFloat, ParameterInt}}: List of parameter objects

Nested Dot Access

x_left_strong = model.geometry.parameters.litho_strong_zones.x_left_strong.value

x_right_strong = model.geometry.parameters.litho_strong_zones.x_right_strong.value

iuse_strong_zones = model.geometry.parameters.litho_strong_zones.iuse_strong_zones.value

Constructor

LithoStrongZones()

Create a new LithoStrongZones parameter group with default values.

Returns

LithoStrongZones: New LithoStrongZones parameter group with initialized default values

EarthBox.ModelDataContainer.GeometryContainer.ParameterCollection.PlumeGroup.Plume — Type

Plume <: AbstractParameterGroup

Parameter group for mantle plume geometry parameters.

Fields

plume_radius::ParameterFloat: Radius in meters of the mantle plume

plume_center_x::ParameterFloat: x-coordinate in meters of the plume center

plume_center_y::ParameterFloat: y-coordinate in meters of the plume center

plume_head_thick::ParameterFloat: Thickness in meters of the plume head

delta_temperature_plume::ParameterFloat: Temperature difference of the plume relative to background mantle

iuse_plume::ParameterInt: Flag for using plume (0=off, 1=on)

plume_start_time::ParameterFloat: Time in Myr when plume starts

iplume_state::ParameterInt: State of the plume: 0 - not initiated; 1 = initiated

obj_list::Vector{Union{ParameterFloat, ParameterInt}}: List of parameter objects

Nested Dot Access

plume_radius = model.geometry.parameters.plume.plume_radius.value

plume_center_x = model.geometry.parameters.plume.plume_center_x.value

plume_center_y = model.geometry.parameters.plume.plume_center_y.value

plume_head_thick = model.geometry.parameters.plume.plume_head_thick.value

delta_temperature_plume = model.geometry.parameters.plume.delta_temperature_plume.value

iuse_plume = model.geometry.parameters.plume.iuse_plume.value

plume_start_time = model.geometry.parameters.plume.plume_start_time.value

iplume_state = model.geometry.parameters.plume.iplume_state.value

Constructor

Plume()

Create a new Plume parameter group with default values.

Returns

Plume: New Plume parameter group with initialized default values

EarthBox.ModelDataContainer.GeometryContainer.ParameterCollection.CrustalHoleGroup.CrustalHole — Type

CrustalHole <: AbstractParameterGroup

Parameter group for crustal hole geometry parameters.

Fields

xhole_start::ParameterFloat: x-coordinate in meters of the start of the crustal hole

xhole_middle::ParameterFloat: x-coordinate in meters of the middle of the crustal hole

xhole_end::ParameterFloat: x-coordinate in meters of the end of the crustal hole

xhole_depth::ParameterFloat: Depth in meters of the crustal hole

obj_list::Vector{Union{[ParameterFloat](@ref)}}: List of parameter objects

Nested Dot Access

xhole_start = model.geometry.parameters.crustal_hole.xhole_start.value

xhole_middle = model.geometry.parameters.crustal_hole.xhole_middle.value

xhole_end = model.geometry.parameters.crustal_hole.xhole_end.value

xhole_depth = model.geometry.parameters.crustal_hole.xhole_depth.value

Constructor

CrustalHole()

Create a new CrustalHole parameter group with default values.

Returns

CrustalHole: New CrustalHole parameter group with initialized default values

EarthBox.ModelDataContainer.GeometryContainer.ParameterCollection.RayleighTaylorGroup.RayleighTaylor — Type

RayleighTaylor <: AbstractParameterGroup

Parameter group for Rayleigh-Taylor instability geometry parameters.

Fields

depth_interface_h1::ParameterFloat: Depth in meters of the interface for Rayleigh-Taylor instability

wave_length_lambda::ParameterFloat: Wavelength in meters of the perturbation for Rayleigh-Taylor instability

amplitude_initial::ParameterFloat: Initial amplitude in meters of the perturbation for Rayleigh-Taylor instability

obj_list::Vector{ParameterFloat}: List of parameter objects

Nested Dot Access

depth_interface_h1 = model.geometry.parameters.rayleigh_taylor.depth_interface_h1.value

wave_length_lambda = model.geometry.parameters.rayleigh_taylor.wave_length_lambda.value

amplitude_initial = model.geometry.parameters.rayleigh_taylor.amplitude_initial.value

Constructor

RayleighTaylor()

Create a new RayleighTaylor parameter group with default values.

Returns

RayleighTaylor: New RayleighTaylor parameter group with initialized default values

EarthBox.ModelDataContainer.GeometryContainer.ParameterCollection.MobileWallGroup.MobileWall — Type

MobileWall <: AbstractParameterGroup

Parameter group for mobile wall geometry parameters.

Fields

x_left_mobile_wall::ParameterFloat: X-location in meters of left-edge of mobile wall used in sandbox extension and shortening models

x_right_mobile_wall::ParameterFloat: X-location in meters of right-edge of mobile wall used in sandbox extension and shortening models

y_top_mobile_wall::ParameterFloat: Y-location in meters of top-edge of mobile wall used in sandbox extension and shortening models

y_bottom_mobile_wall::ParameterFloat: Y-location in meters of bottom-edge of mobile wall used in sandbox extension and shortening models

plate_extension_width::ParameterFloat: Width in meters of plate extension used in sandbox extension models

plate_extension_thickness::ParameterFloat: Thickness in meters of plate extension used in sandbox extension models

obj_list::Vector{ParameterFloat}: List of parameter objects

Nested Dot Access

x_left_mobile_wall = model.geometry.parameters.mobile_wall.x_left_mobile_wall.value

x_right_mobile_wall = model.geometry.parameters.mobile_wall.x_right_mobile_wall.value

y_top_mobile_wall = model.geometry.parameters.mobile_wall.y_top_mobile_wall.value

y_bottom_mobile_wall = model.geometry.parameters.mobile_wall.y_bottom_mobile_wall.value

plate_extension_width = model.geometry.parameters.mobile_wall.plate_extension_width.value

plate_extension_thickness = model.geometry.parameters.mobile_wall.plate_extension_thickness.value

Constructor

MobileWall()

Create a new MobileWall parameter group with default values.

Returns

MobileWall: New MobileWall parameter group with initialized default values

EarthBox.ModelDataContainer.GeometryContainer.ParameterCollection.SandboxGroup.Sandbox — Type

Sandbox <: AbstractParameterGroup

Parameter group for sandbox model geometry parameters.

Fields

nsand_layers::ParameterInt: Number of sand layers used in sandbox extension and shortening models

y_sand_air_interface::ParameterFloat: Y-location in meters of interface between sand and air along the left boundary used in sandbox extension and shortening models

y_top_microbeads::ParameterFloat: Y-location in meters of top of microbeads used in sandbox shortening models

y_bottom_microbeads::ParameterFloat: Y-location in meters of bottom of microbeads used in sandbox shortening models

x_left_ramp::ParameterFloat: X-location in meters of left-edge of ramp used in sandbox shortening models

x_right_ramp::ParameterFloat: X-location in meters of right-edge of ramp used in sandbox shortening models

ramp_dip_deg::ParameterFloat: Dip of ramp in degrees used in sandbox shortening models used in sandbox shortening models

pdms_layer_width::ParameterFloat: Width in meters of PDMS (polydimethylsiloxane) layer located in the central bottom part of model used in sandbox extension models

pdms_layer_thickness::ParameterFloat: Thickness in meters of PDMS (polydimethylsiloxane) layer used in sandbox extension models

obj_list::Vector{Union{ParameterFloat, ParameterInt}}`: List of parameter objects

Nested Dot Access

nsand_layers = model.geometry.parameters.sandbox.nsand_layers.value

y_sand_air_interface = model.geometry.parameters.sandbox.y_sand_air_interface.value

y_top_microbeads = model.geometry.parameters.sandbox.y_top_microbeads.value

y_bottom_microbeads = model.geometry.parameters.sandbox.y_bottom_microbeads.value

x_left_ramp = model.geometry.parameters.sandbox.x_left_ramp.value

x_right_ramp = model.geometry.parameters.sandbox.x_right_ramp.value

ramp_dip_deg = model.geometry.parameters.sandbox.ramp_dip_deg.value

pdms_layer_width = model.geometry.parameters.sandbox.pdms_layer_width.value

pdms_layer_thickness = model.geometry.parameters.sandbox.pdms_layer_thickness.value

Constructor

Sandbox()

Create a new Sandbox parameter group with default values.

Returns

Sandbox: New Sandbox parameter group with initialized default values

EarthBox.ModelDataContainer.GeometryContainer.ParameterCollection.InternalVelocityZoneGroup.InternalVelocityZone — Type

InternalVelocityZone <: AbstractParameterGroup

Parameter group for internal velocity zone geometry parameters.

Fields

xindex_vx_internal::ParameterInt: X-index for internal vx velocity

yindex_min_vx_internal::ParameterInt: Minimum y-index for internal vx velocity

yindex_max_vx_internal::ParameterInt: Maximum y-index for internal vx velocity

xindex_vy_internal::ParameterInt: X-index for internal vy velocity

yindex_min_vy_internal::ParameterInt: Minimum y-index for internal vy velocity

yindex_max_vy_internal::ParameterInt: Maximum y-index for internal vy velocity

x_vx_internal::ParameterFloat: X-coordinate in meters for internal vx velocity

y_min_vx_internal::ParameterFloat: Minimum y-coordinate in meters for internal vx velocity

y_max_vx_internal::ParameterFloat: Maximum y-coordinate in meters for internal vx velocity

x_vy_internal::ParameterFloat: X-coordinate in meters for internal vy velocity

y_min_vy_internal::ParameterFloat: Minimum y-coordinate in meters for internal vy velocity

y_max_vy_internal::ParameterFloat: Maximum y-coordinate in meters for internal vy velocity

obj_list::Vector{Union{ParameterFloat, ParameterInt}}: List of parameter objects

Nested Dot Access

xindex_vx_internal = model.geometry.parameters.internal_velocity_zone.xindex_vx_internal.value

yindex_min_vx_internal = model.geometry.parameters.internal_velocity_zone.yindex_min_vx_internal.value

yindex_max_vx_internal = model.geometry.parameters.internal_velocity_zone.yindex_max_vx_internal.value

xindex_vy_internal = model.geometry.parameters.internal_velocity_zone.xindex_vy_internal.value

yindex_min_vy_internal = model.geometry.parameters.internal_velocity_zone.yindex_min_vy_internal.value

yindex_max_vy_internal = model.geometry.parameters.internal_velocity_zone.yindex_max_vy_internal.value

x_vx_internal = model.geometry.parameters.internal_velocity_zone.x_vx_internal.value

y_min_vx_internal = model.geometry.parameters.internal_velocity_zone.y_min_vx_internal.value

y_max_vx_internal = model.geometry.parameters.internal_velocity_zone.y_max_vx_internal.value

x_vy_internal = model.geometry.parameters.internal_velocity_zone.x_vy_internal.value

y_min_vy_internal = model.geometry.parameters.internal_velocity_zone.y_min_vy_internal.value

y_max_vy_internal = model.geometry.parameters.internal_velocity_zone.y_max_vy_internal.value

Constructor

InternalVelocityZone()

Create a new InternalVelocityZone parameter group with default values.

Returns

InternalVelocityZone: New InternalVelocityZone parameter group with initialized default values

EarthBox.ModelDataContainer.GeometryContainer.ParameterCollection.FractureZoneGroup.FractureZone — Type

FractureZone <: AbstractParameterGroup

Parameter group for fracture zone geometry parameters.

Fields

sediment_thickness::ParameterFloat: Thickness in meters of sediment layer

basaltic_oceanic_crust_thickness::ParameterFloat: Thickness in meters of basaltic oceanic crust layer

gabbroic_oceanic_crust_thickness::ParameterFloat: Thickness in meters of gabbroic oceanic crust layer

thickness_of_younger_lithosphere::ParameterFloat: Thickness in meters of younger lithosphere

thickness_of_older_lithosphere::ParameterFloat: Thickness in meters of older lithosphere

thickness_of_weak_lithosphere::ParameterFloat: Thickness in meters of weak lithosphere

x_fracture_zone_start::ParameterFloat: Start of fracture zone in x-direction in meters

x_fracture_zone_end::ParameterFloat: End of fracture zone in x-direction in meters

obj_list::Vector{Union{ParameterFloat, ParameterInt}}: List of parameter objects

Nested Dot Access

sediment_thickness = model.geometry.parameters.fracture_zone.sediment_thickness.value

basaltic_oceanic_crust_thickness = model.geometry.parameters.fracture_zone.basaltic_oceanic_crust_thickness.value

gabbroic_oceanic_crust_thickness = model.geometry.parameters.fracture_zone.gabbroic_oceanic_crust_thickness.value

thickness_of_younger_lithosphere = model.geometry.parameters.fracture_zone.thickness_of_younger_lithosphere.value

thickness_of_older_lithosphere = model.geometry.parameters.fracture_zone.thickness_of_older_lithosphere.value

thickness_of_weak_lithosphere = model.geometry.parameters.fracture_zone.thickness_of_weak_lithosphere.value

x_fracture_zone_start = model.geometry.parameters.fracture_zone.x_fracture_zone_start.value

x_fracture_zone_end = model.geometry.parameters.fracture_zone.x_fracture_zone_end.value

Constructor

FractureZone()

Create a new FractureZone parameter group with default values.

Returns

FractureZone: New FractureZone parameter group with initialized default values