Grids

Parameters

GridOptions

GridGeometry

GridRefinement

Arrays

BasicGrid

StaggeredVyGrid

StaggeredVxGrid

PressureGrid

EarthBox.ModelDataContainer.Grids2dContainer.Grids — Type

Grids <: CollectionContainer

Data structure containing parameter and array objects associated with 2D staggered grids.

Fields

parameters::Parameters: Parameter groups for basic and staggered grids

arrays::Arrays: Array groups containing coordinates and spacing for basic and staggered

Constructor

Grids(ynum::Int, xnum::Int, ysize::Float64, xsize::Float64)::Grids

Create a new Grids collection with the given grid dimensions.

Arguments

ynum::Int: Number of grid points in y-direction

xnum::Int: Number of grid points in x-direction

ysize::Float64: Domain size in y-direction in meters

xsize::Float64: Domain size in x-direction in meters

EarthBox.ModelDataContainer.Grids2dContainer.ParameterCollection.Parameters — Type

Parameters <: AbstractParameterCollection

Parameter collection for 2D staggered grids.

Fields

grid_options::GridOptions: Grid type options

geometry::GridGeometry: Dimensions and resolution of basic grids

refinement::GridRefinement: T-type grid refinement parameters

Constructor

Parameters(ynum::Int, xnum::Int, ysize::Float64, xsize::Float64)

Create a new Parameters collection with the given grid dimensions.

Arguments

ynum::Int: Number of grid points in y-direction

xnum::Int: Number of grid points in x-direction

ysize::Float64: Domain size in y-direction in meters

xsize::Float64: Domain size in x-direction in meters

EarthBox.ModelDataContainer.Grids2dContainer.ParameterCollection.GridOptionsGroup.GridOptions — Type

GridOptions <: AbstractParameterGroup

Parameter group for grid options.

Fields

itype_grid::ParameterInt: Grid option ID

stype_grid::ParameterStr: Grid option name

obj_list::Vector{ParameterInt}: List of parameter objects

Nested Dot Access

itype_grid = model.grids.parameters.grid_options.itype_grid.value

stype_grid = model.grids.parameters.grid_options.stype_grid.value

Constructor

GridOptions()

Create a new GridOptions parameter group with default values.

Returns

GridOptions: New GridOptions parameter group with initialized values

EarthBox.ModelDataContainer.Grids2dContainer.ParameterCollection.GridGeometryGroup.GridGeometry — Type

GridGeometry <: AbstractParameterGroup

Parameter group for basic grid geometry parameters.

Fields

ynum::ParameterInt: Number of basic grid points in y-direction

xnum::ParameterInt: Number of basic grid points in x-direction

ysize::ParameterFloat: Height of model in meters

xsize::ParameterFloat: Width of model in meters

ystpavg::ParameterFloat: Average spacing of basic grid in y-direction in meters

xstpavg::ParameterFloat: Average spacing of basic grid in x-direction in meters

ymin::ParameterFloat: Minimum y-coordinate of model domain in meters

ymax::ParameterFloat: Maximum y-coordinate of model domain in meters

xmin::ParameterFloat: Minimum x-coordinate of model domain in meters

xmax::ParameterFloat: Maximum x-coordinate of model domain in meters

xsize_start::ParameterFloat: Initial width of model in meters

obj_list::Vector{Union{ParameterFloat, ParameterInt}}: List of parameter objects

Nested Dot Access

ynum = model.grids.parameters.geometry.ynum.value

xnum = model.grids.parameters.geometry.xnum.value

ysize = model.grids.parameters.geometry.ysize.value

xsize = model.grids.parameters.geometry.xsize.value

ystpavg = model.grids.parameters.geometry.ystpavg.value

xstpavg = model.grids.parameters.geometry.xstpavg.value

ymin = model.grids.parameters.geometry.ymin.value

ymax = model.grids.parameters.geometry.ymax.value

xmin = model.grids.parameters.geometry.xmin.value

xmax = model.grids.parameters.geometry.xmax.value

xsize_start = model.grids.parameters.geometry.xsize_start.value

Constructor

GridGeometry(ynum::Int, xnum::Int, ysize::Float64, xsize::Float64)

Create a new GridGeometry parameter group with the given grid dimensions.

Arguments

ynum::Int: Number of grid points in y-direction

xnum::Int: Number of grid points in x-direction

ysize::Float64: Domain size in y-direction in meters

xsize::Float64: Domain size in x-direction in meters

Returns

GridGeometry: New GridGeometry parameter group with initialized values

EarthBox.ModelDataContainer.Grids2dContainer.ParameterCollection.GridRefinementGroup.GridRefinement — Type

GridRefinement <: AbstractParameterGroup

Parameter group for grid refinement parameters.

Fields

dx_highres::ParameterFloat: Constant horizontal grid resolution in high-resolution area in meters

dx_lowres::ParameterFloat: Average horizontal grid resolution in low-resolution area in meters

xo_highres::ParameterFloat: x-location of first node of high-resolution area in meters

ixo_highres::ParameterInt: x-index of first node of high-resolution area

xf_highres::ParameterFloat: x-location of last node of high-resolution area in meters

dy_highres::ParameterFloat: Constant vertical grid resolution in high-resolution area in meters

dy_lowres::ParameterFloat: Average vertical grid resolution in low-resolution area in meters

yf_highres::ParameterFloat: y-location of last node of high-resolution area in meters

iuse_trench::ParameterInt: Flag to use trench location for refinement

trench_location::ParameterFloat: x-location of trench in meters

iuse_refinement_delay::ParameterInt: Flag to delay grid refinement

refinement_time::ParameterFloat: Time to start grid refinement in Myr

refinement_flag::ParameterInt: Flag indicating if grid is currently refined

iuse_refinement_gap::ParameterInt: Flag to temporarily disable refinement

refinement_gap_start_time::ParameterFloat: Time to start refinement gap in Myr

refinement_gap_end_time::ParameterFloat: Time to end refinement gap in Myr

obj_list::Vector{Union{ParameterFloat, ParameterInt}}: List of parameter objects

Nested Dot Access

dx_highres = model.grids.parameters.refinement.dx_highres.value

dx_lowres = model.grids.parameters.refinement.dx_lowres.value

xo_highres = model.grids.parameters.refinement.xo_highres.value

ixo_highres = model.grids.parameters.refinement.ixo_highres.value

xf_highres = model.grids.parameters.refinement.xf_highres.value

dy_highres = model.grids.parameters.refinement.dy_highres.value

dy_lowres = model.grids.parameters.refinement.dy_lowres.value

yf_highres = model.grids.parameters.refinement.yf_highres.value

iuse_trench = model.grids.parameters.refinement.iuse_trench.value

trench_location = model.grids.parameters.refinement.trench_location.value

iuse_refinement_delay = model.grids.parameters.refinement.iuse_refinement_delay.value

refinement_time = model.grids.parameters.refinement.refinement_time.value

refinement_flag = model.grids.parameters.refinement.refinement_flag.value

iuse_refinement_gap = model.grids.parameters.refinement.iuse_refinement_gap.value

refinement_gap_start_time = model.grids.parameters.refinement.refinement_gap_start_time.value

refinement_gap_end_time = model.grids.parameters.refinement.refinement_gap_end_time.value

Constructor

GridRefinement()

Create a new GridRefinement parameter group with default values.

Returns

GridRefinement: New GridRefinement parameter group with initialized values

EarthBox.ModelDataContainer.Grids2dContainer.ArrayCollection.Arrays — Type

Arrays <: AbstractArrayCollection

Data structure containing array groups for basic and staggered grids.

Fields

basic::BasicGrid: Coordinate and nodal spacing arrays for basic grid

staggered_vy::StaggeredVyGrid: Coordinate and nodal spacing arrays for staggered Vy grid

staggered_vx::StaggeredVxGrid: Coordinate and nodal spacing arrays for staggered Vx grid

pressure::PressureGrid: Coordinate and nodal spacing arrays for pressure nodes

Constructor

Arrays(ynum::Int, xnum::Int)::Arrays

Create a new Arrays collection with the given grid dimensions.

Arguments

ynum::Int: Number of grid points in y-direction

xnum::Int: Number of grid points in x-direction

EarthBox.ModelDataContainer.Grids2dContainer.ArrayCollection.BasicGridGroup.BasicGrid — Type

BasicGrid <: AbstractArrayGroup

Array group containing coordinates and spacing for basic grid.

Fields

gridx_b::GridArray1DState: (xnum) : x-coordinates of basic grid nodes in meters.

xstp_b::GridArray1DState: (xnum - 1) : Width of cells in x-direction for basic grid in meters.

gridy_b::GridArray1DState: (ynum) : y-coordinates of basic grid nodes in meters.

ystp_b::GridArray1DState: (ynum - 1) : Width of cells in y-direction for basic grid in meters.

Nested Dot Access

gridx_b = model.grids.arrays.basic.gridx_b.array

xstp_b = model.grids.arrays.basic.xstp_b.array

gridy_b = model.grids.arrays.basic.gridy_b.array

ystp_b = model.grids.arrays.basic.ystp_b.array

Constructor

BasicGrid(ynum::Int, xnum::Int)

Create a new BasicGrid array group with the given grid dimensions.

Arguments

ynum::Int: Number of grid points in y-direction

xnum::Int: Number of grid points in x-direction

Returns

BasicGrid: New BasicGrid array group with initialized arrays

EarthBox.ModelDataContainer.Grids2dContainer.ArrayCollection.StaggeredVyGridGroup.StaggeredVyGrid — Type

StaggeredVyGrid <: AbstractArrayGroup

Array group containing coordinates and spacing for staggered Vy grid.

Fields

gridx_vy::GridArray1DState: (xnum + 1) : x-locations of staggered Vy grid nodes including ghost nodes in meters.

xstp_vy::GridArray1DState: (xnum) : Width of staggered vy grid cells in x-direction in meters.

Nested Dot Access

gridx_vy = model.grids.arrays.staggered_vy.gridx_vy.array

xstp_vy = model.grids.arrays.staggered_vy.xstp_vy.array

Constructor

StaggeredVyGrid(xnum::Int)

Create a new StaggeredVyGrid array group with the given grid dimensions.

Arguments

xnum::Int: Number of grid points in x-direction

Returns

StaggeredVyGrid: New StaggeredVyGrid array group with initialized arrays

EarthBox.ModelDataContainer.Grids2dContainer.ArrayCollection.StaggeredVxGridGroup.StaggeredVxGrid — Type

StaggeredVxGrid <: AbstractArrayGroup

Array group containing coordinates and spacing for staggered Vx grid.

Fields

gridy_vx::GridArray1DState: (ynum + 1) : y-locations of staggered Vx grid nodes including ghost nodes in meters.

ystp_vx::GridArray1DState: (ynum) : Width of staggered vx grid cells in y-direction in meters.

Nested Dot Access

gridy_vx = model.grids.arrays.staggered_vx.gridy_vx.array

ystp_vx = model.grids.arrays.staggered_vx.ystp_vx.array

Constructor

StaggeredVxGrid(ynum::Int)

Create a new StaggeredVxGrid array group with the given grid dimensions.

Arguments

ynum::Int: Number of grid points in y-direction

Returns

StaggeredVxGrid: New StaggeredVxGrid array group with initialized arrays

EarthBox.ModelDataContainer.Grids2dContainer.ArrayCollection.PressureGridGroup.PressureGrid — Type

PressureGrid <: AbstractArrayGroup

Array group containing coordinates and spacing for pressure grid.

Fields

gridy_pr::GridArray1DState: (ynum - 1) : y-locations of pressure grid nodes in meters.

ystp_pr::GridArray1DState: (ynum - 2) : Width of pressure cells in y-direction in meters.

gridx_pr::GridArray1DState: (xnum - 1) : x-locations of pressure grid nodes in meters.

xstp_pr::GridArray1DState: (xnum - 2) : Width of pressure cells in x-direction in meters.

Nested Dot Access

gridy_pr = model.grids.arrays.pressure.gridy_pr.array

ystp_pr = model.grids.arrays.pressure.ystp_pr.array

gridx_pr = model.grids.arrays.pressure.gridx_pr.array

xstp_pr = model.grids.arrays.pressure.xstp_pr.array

Constructor

PressureGrid(ynum::Int, xnum::Int)

Create a new PressureGrid array group with the given grid dimensions.

Arguments

ynum::Int: Number of grid points in y-direction

xnum::Int: Number of grid points in x-direction

Returns

PressureGrid: New PressureGrid array group with initialized arrays