Topography

Parameters

ModelOptions

TopoGrid

DepoAndErosionRates

DownhillDiffusion

SeaLevel

ReferenceLithosphere

Arrays

EarthBox.ModelDataContainer.TopographyContainer.Topography — Type

Topography <: CollectionContainer

Data structure containing parameter and array objects for topography evolution.

Fields

parameters::Parameters: Parameter groups for topography configuration

arrays::Arrays: Array groups for topographic grid data

Constructor

Topography()::Topography

Create a new Topography collection with default parameters.

EarthBox.ModelDataContainer.TopographyContainer.ParameterCollection.Parameters — Type

Parameters <: AbstractParameterCollection

Parameter collection for topography evolution model.

Fields

model_options::ModelOptions: Topography model on/off switches and options

topo_grid::TopoGrid: Topography grid parameters

depo_and_erosion_rates::DepoAndErosionRates: Deposition and erosion rates

downhill_diffusion::DownhillDiffusion: Downhill diffusion parameters

sealevel::SeaLevel: Sea level parameters

reference_lithosphere::ReferenceLithosphere: Reference lithosphere parameters

Constructor

Parameters()

Create a new Parameters collection with default topography model values.

EarthBox.ModelDataContainer.TopographyContainer.ParameterCollection.ModelOptionsGroup.ModelOptions — Type

ModelOptions <: AbstractParameterGroup

Parameter group for topography model options.

Fields

iuse_topo::ParameterInt: Integer flag used to enable topography model: 0 = off; 1 = on

itype_topo::ParameterInt: ID for topography node advection option: 0 = Runge Kutta with interpolation; 1 = antidiffusion

stype_topo::ParameterStr: Topography advection option name

iuse_downhill_diffusion::ParameterInt: Integer flag that activates sediment transport model with downhill diffusion (0 = off, 1 = on). If option is set to off uniform erosion and deposition rates will be used

iuse_salt_deposition::ParameterInt: Integer flag that activates salt deposition model: 0 = off; 1 = on

obj_list::Vector{Union{ParameterFloat, ParameterInt, ParameterStr}}: List of parameter objects

Nested Dot Access

iuse_topo = model.topography.parameters.model_options.iuse_topo.value

itype_topo = model.topography.parameters.model_options.itype_topo.value

stype_topo = model.topography.parameters.model_options.stype_topo.value

iuse_downhill_diffusion = model.topography.parameters.model_options.iuse_downhill_diffusion.value

iuse_salt_deposition = model.topography.parameters.model_options.iuse_salt_deposition.value

Constructor

ModelOptions()

Returns

ModelOptions: New ModelOptions parameter group with initialized values

EarthBox.ModelDataContainer.TopographyContainer.ParameterCollection.TopoGridGroup.TopoGrid — Type

TopoGrid <: AbstractParameterGroup

Parameter group for topography grid parameters.

Fields

topo_xsize::ParameterFloat: Topography model size in horizontal direction in meters

toponum::ParameterInt: Number of nodes in the x-direction of the topography model

dx_topo::ParameterFloat: Topography grid step size in meters

nsmooth_top_bottom::ParameterInt: Number of topography nodes that will be used to calculate a running average for the calculated y-coordinate of the top and bottom of marker layers

marker_search_factor::ParameterFloat: Search factor used to calculate the search radius for determining the shallowest and deepest markers associated with a given topography node

obj_list::Vector{Union{ParameterFloat, ParameterInt}}: List of parameter objects

Nested Dot Access

topo_xsize = model.topography.parameters.topo_grid.topo_xsize.value

toponum = model.topography.parameters.topo_grid.toponum.value

dx_topo = model.topography.parameters.topo_grid.dx_topo.value

nsmooth_top_bottom = model.topography.parameters.topo_grid.nsmooth_top_bottom.value

marker_search_factor = model.topography.parameters.topo_grid.marker_search_factor.value

Constructor

TopoGrid()

Returns

TopoGrid: New TopoGrid parameter group with initialized values

EarthBox.ModelDataContainer.TopographyContainer.ParameterCollection.DepoAndErosionRatesGroup.DepoAndErosionRates — Type

DepoAndErosionRates <: AbstractParameterGroup

Parameter group for deposition and erosion rates.

Fields

erosion_rate::ParameterFloat: Erosion rate above water level in m/s

sedimentation_rate::ParameterFloat: Sedimentation rate below water level in m/s

pelagic_sedimentation_rate::ParameterFloat: Pelagic sedimentation rate (m/s). Typical value used in the literature are 0.3 mm/yr (syn-rift) to 0.01 mm/yr (post-rift) (e.g. Perez-Gussinye et al., 2020)

pelagic_sedimentation_rate_reduction_factor::ParameterFloat: The pelagic sedimentation rate is divided by this factor after the specified pelagic sedimentation rate reduction time.

pelagic_sedimentation_rate_reduction_time::ParameterFloat: Time in Myr after which the pelagic sedimentation rate is reduced by the pelagicsedimentationratereductionfactor.

salt_deposition_rate::ParameterFloat: Salt deposition rate in m/s

salt_start_time::ParameterFloat: Salt start time in Myr

salt_end_time::ParameterFloat: Salt end time in Myr

obj_list::Vector{Union{ParameterFloat, ParameterInt}}: List of parameter objects

Nested Dot Access

erosion_rate = model.topography.parameters.depo_and_erosion_rates.erosion_rate.value

sedimentation_rate = model.topography.parameters.depo_and_erosion_rates.sedimentation_rate.value

pelagic_sedimentation_rate = model.topography.parameters.depo_and_erosion_rates.pelagic_sedimentation_rate.value

pelagic_sedimentation_rate_reduction_factor = model.topography.parameters.depo_and_erosion_rates.pelagic_sedimentation_rate_reduction_factor.value

pelagic_sedimentation_rate_reduction_time = model.topography.parameters.depo_and_erosion_rates.pelagic_sedimentation_rate_reduction_time.value

salt_deposition_rate = model.topography.parameters.depo_and_erosion_rates.salt_deposition_rate.value

salt_start_time = model.topography.parameters.depo_and_erosion_rates.salt_start_time.value

salt_end_time = model.topography.parameters.depo_and_erosion_rates.salt_end_time.value

Constructor

DepoAndErosionRates()

Returns

DepoAndErosionRates: New DepoAndErosionRates parameter group with initialized values

EarthBox.ModelDataContainer.TopographyContainer.ParameterCollection.DownhillDiffusionGroup.DownhillDiffusion — Type

DownhillDiffusion <: AbstractParameterGroup

Parameter group for downhill diffusion parameters.

Fields

downhill_diff_elev_max::ParameterFloat: d2Yt - max elevation

transport_length::ParameterFloat: Transport length scale in meters

topo_diff_coef::ParameterFloat: Topography diffusion coefficient in m^2/s

subaerial_slope_diffusivity::ParameterFloat: Subaerial slope diffusivity in m^2/s. Typical value used in the literature is 0.25 m^2/yr (7.9e-9 m^2/s) (e.g. Andres-Martinez et al., 2019; Armitage et al., 2015)

precipitation_rate::ParameterFloat: Precipitation rate in m/s used to calculate water flux in drainage basins for fluvial transport diffusivity model. Precipitation rate (m/s). Used to calculate water flux in drainage basins. Typical value used in the literature is 1 m/yr (3.2e-8 m/s) (e.g. Andres-Martinez et al.; 2019; Huffman et al., 2009).

subaerial_transport_coefficient::ParameterFloat: Subaerial discharge transport coefficient. Used to calculate effective subaerial fluvial transport diffusivity that includes slope diffusivity, precipitation rate and downstream distances. Typical values used in the literature are 1e-4 (low transport) to 1e-2 (high transport) (e.g. Andres-Martinez et al., 2019; Armitage et al., 2015).

submarine_slope_diffusivity::ParameterFloat: Maximum submarine slope diffusivity in m^2/s used in diffusivity model that decays exponentially with water depth. Typical value used in the literature is 100 m^2/yr (3.2e-9 m^2/s) (e.g. Andres-Martinez et al., 2019; Kaufman et al., 1991).

submarine_diffusion_decay_depth::ParameterFloat: Submarine diffusion decay depth (m). Typical value used in the literature is 1000-2000 m (e.g. Andres-Martinez et al., 2019; Kaufmanet al., 1991; Perez-Gussinye et al., 2020).

transport_timestep::ParameterFloat: Transport timestep (s). Typical value used in the literature is 1000 years (3.1536e10 s) (e.g. Andres-Martinez et al., 2019).

number_of_transport_timesteps_per_model_timestep::ParameterInt: Number of transport time steps per model time step.

iuse_compaction_correction::ParameterInt: Integer flag activating compaction correction during sediment transport where newly deposited sediment thickness and pre-existing sediment thickness are corrected for compaction.

obj_list::Vector{Union{ParameterFloat, ParameterInt}}: List of parameter objects

Nested Dot Access

downhill_diff_elev_max = model.topography.parameters.downhill_diffusion.downhill_diff_elev_max.value

transport_length = model.topography.parameters.downhill_diffusion.transport_length.value

topo_diff_coef = model.topography.parameters.downhill_diffusion.topo_diff_coef.value

subaerial_slope_diffusivity = model.topography.parameters.downhill_diffusion.subaerial_slope_diffusivity.value

precipitation_rate = model.topography.parameters.downhill_diffusion.precipitation_rate.value

subaerial_transport_coefficient = model.topography.parameters.downhill_diffusion.subaerial_transport_coefficient.value

submarine_slope_diffusivity = model.topography.parameters.downhill_diffusion.submarine_slope_diffusivity.value

submarine_diffusion_decay_depth = model.topography.parameters.downhill_diffusion.submarine_diffusion_decay_depth.value

transport_timestep = model.topography.parameters.downhill_diffusion.transport_timestep.value

iuse_compaction_correction = model.topography.parameters.downhill_diffusion.iuse_compaction_correction.value

Constructor

DownhillDiffusion()

Returns

DownhillDiffusion: New DownhillDiffusion parameter group with initialized values

EarthBox.ModelDataContainer.TopographyContainer.ParameterCollection.SeaLevelGroup.SeaLevel — Type

SeaLevel <: AbstractParameterGroup

Parameter group for sea level parameters.

Fields

itype_sealevel::ParameterInt: Integer flag controlling the sealevel model

stype_sealevel::ParameterStr: Sealevel option name

y_water_ini::ParameterFloat: Initial and current seawater level in meters

base_level_shift::ParameterFloat: Shift in base level in meters. Positive values shift the base level down in the +y direction.

base_level_shift_end_time::ParameterFloat: End time for base level shift in Myr

y_sealevel::ParameterFloat: Seawater level y-coordinate in meters

obj_list::Vector{Union{ParameterFloat, ParameterInt, ParameterStr}}: List of parameter objects

Nested Dot Access

itype_sealevel = model.topography.parameters.sealevel.itype_sealevel.value

stype_sealevel = model.topography.parameters.sealevel.stype_sealevel.value

y_water_ini = model.topography.parameters.sealevel.y_water_ini.value

base_level_shift = model.topography.parameters.sealevel.base_level_shift.value

base_level_shift_end_time = model.topography.parameters.sealevel.base_level_shift_end_time.value

y_sealevel = model.topography.parameters.sealevel.y_sealevel.value

Constructor

SeaLevel()

Returns

SeaLevel: New SeaLevel parameter group with initialized values

EarthBox.ModelDataContainer.TopographyContainer.ParameterCollection.ReferenceLithosphereGroup.ReferenceLithosphere — Type

ReferenceLithosphere <: AbstractParameterGroup

Parameter group for reference lithosphere parameters.

Fields

thickness_upper_continental_crust_ref::ParameterFloat: Thickness of upper continental crust of the reference lithosphere in meters

thickness_lower_continental_crust_ref::ParameterFloat: Thickness of lower continental crust of the reference lithosphere in meters

thickness_lithosphere_ref::ParameterFloat: Thickness of lithosphere for the reference lithosphere in meters

gridy_spacing_ref::ParameterFloat: Spacing of grid cells in y-direction for the reference lithosphere in meters

temperature_top_ref::ParameterFloat: Temperature at top of lithosphere for the reference lithosphere in Celsius

temperature_moho_ref::ParameterFloat: Temperature at Moho of the reference lithosphere in Celsius

temperature_base_lith_ref::ParameterFloat: Temperature at base of lithosphere for the reference lithosphere in Celsius

adiabatic_gradient_ref::ParameterFloat: Adiabatic gradient below the reference lithosphere in K/km

iuse_linear_segments::ParameterInt: Integer flag that controls the type of temperature profile used in the reference lithosphere model. If iuse_linear_segments = 1, a temperature profile with four linear segments is used that is controlled by the following user defined parameters: temperature_top_ref, temperature_moho_ref, and temperature_base_lith_ref. If iuse_linear_segments = 0, the temperature profile is calculated using an analytical 3-layer model defined during marker temperature. See the AnalyticalThreeLayer model description at MarkerTemperature). Default is 1.

obj_list::Vector{Union{ParameterFloat, ParameterInt}}: List of parameter objects

Nested Dot Access

thickness_upper_continental_crust_ref = model.topography.parameters.reference_lithosphere.thickness_upper_continental_crust_ref.value

thickness_lower_continental_crust_ref = model.topography.parameters.reference_lithosphere.thickness_lower_continental_crust_ref.value

thickness_lithosphere_ref = model.topography.parameters.reference_lithosphere.thickness_lithosphere_ref.value

gridy_spacing_ref = model.topography.parameters.reference_lithosphere.gridy_spacing_ref.value

temperature_top_ref = model.topography.parameters.reference_lithosphere.temperature_top_ref.value

temperature_moho_ref = model.topography.parameters.reference_lithosphere.temperature_moho_ref.value

temperature_base_lith_ref = model.topography.parameters.reference_lithosphere.temperature_base_lith_ref.value

adiabatic_gradient_ref = model.topography.parameters.reference_lithosphere.adiabatic_gradient_ref.value

iuse_linear_segments = model.topography.parameters.reference_lithosphere.iuse_linear_segments.value

Constructor

ReferenceLithosphere()

Returns

ReferenceLithosphere: New ReferenceLithosphere parameter group with initialized values

EarthBox.ModelDataContainer.TopographyContainer.ArrayCollection.Arrays — Type

Arrays <: AbstractArrayCollection

Data structure containing array groups for topography evolution.

Fields

gridt::TopoArray2DState: (7, toponum) Topography grid array

ROW 1: x-coordinates of topography grid nodes.

ROW 2: y-coordinate of topography grid nodes (i.e. elevation).

ROW 3: Updated pre-antidiffusion y-coordinate (i.e. elevation) at topography grid nodes.

ROW 4: x-component of velocity at topography grid nodes.

ROW 5: y-component of velocity at topography grid nodes.

ROW 6: Antidiffusion correction.

ROW 7: Thickness of extrusive material at each node.

Nested Dot Access

model.topography.arrays.gridt.array

Constructor

Arrays(toponum::Int)::Arrays

Create a new Arrays collection with the given topography grid resolution.

Arguments

toponum::Int: Number of topography grid points