ModelData

EarthBox.ModelDataContainer.ModelData — Type

ModelData

Struct containing parameter and array collections for the EarthBox model data and other fields related to the EarthBox model. Each collection container contains parameter and array collections, parameters and arrays respectively, for the various data categories. Parameters and array collections are in turn composed of groups of related parameters and arrays. The general structure of the model data collections container is as follows:

ModelData
├─ collection_container_1
│  ├─ parameters
│  │  ├─ parameter_group_A
│  │  │  └─ parameter_A1
│  │  │  └─ parameter_A2
│  │  └─ parameter_group_B
│  │     ├─ parameter_B1
│  │     ├─ parameter_B2
│  └─ arrays
│     ├─ array_group_C
│     │  └─ array_C1
│     │  └─ array_C2
│     └─ array_group_D
│        └─ array_D1
│        └─ array_D2
├─ collection_container_2
│  ├─ parameters
│  │  ├─ parameter_group_E
│  │  │  └─ parameter_E1
│  │  │  └─ parameter_E2
│  │  └─ parameter_group_F
│  │     ├─ parameter_F1
│  │     ├─ parameter_F2
│  └─ arrays
│     ├─ array_group_G
│     │  └─ array_G1
│     │  └─ array_G2
│     └─ array_group_H
│        └─ array_H1
│        └─ array_H2

For a ModelData object model, the parameter_A1 and array_C1 can be accessed as follows:

model.collection_container_1.parameters.parameter_group_A.parameter_A1

model.collection_container_1.arrays.array_group_C.array_C1

Parameter groups contain parameter objects with attributes like value, unit, and description. Array groups contain array objects with attributes like array, units, and description.

Fields

grids::Grids or Grids3d: Parameter and array collections for basic and staggered grids. See Grids for details.

timestep::Timestep: Parameter collections for model time stepping. See Timestep for details.

geometry::Geometry: Parameter collections for material geometry. See Geometry for details.

bcs::BoundaryConditions: Parameter and array collections for boundary conditions. See BoundaryConditions for details.

materials::Materials: Parameter, array and dictionary collections for material properties. See Materials for details.

markers::Markers: Parameter and array collections for markers. See Markers for details.

interpolation::Interpolation: Parameter and array collections for marker-grid interpolation. See Interpolation for details.

melting::Melting: Parameter and array collections for melting and melt extraction. See Melting for details.

topography::Topography: Parameter and array collections for topography evolution. See Topography for details.

heat_equation::HeatEquation: Parameter and array collections for heat equation. See HeatEquation for details.

stokes_continuity::StokesContinuity: Parameter and array collections for Stokes-continuity solver. See StokesContinuity for details.

gravity: Gravity parameters including gravity components and control flags. See Gravity for details.

benchmarks::Benchmarks: Parameter collection for benchmark parameters. See Benchmarks for details.

conversion::Conversion: Conversion parameters. See Conversion for details.

carbonate::Carbonate: Parameter and array collections for carbonate deposition configuration. See Carbonate for details.

RGAS::ParameterFloat: Gas constant. See ParameterFloat for details on type.

unit_conversion_data::UnitConversionData: Data structure with unit conversion functions and registry.

model_input_file::Union{String, Nothing}: Model input file.

input_parameters::ParametersDictType: Input parameters dictionary.

obj_dict::ObjDictType: Dictionary of EarthBox parameter and array objects where keys are names and values are parameter and array objects.

global_iter_dict::GlobalIterDictType: Dictionary of global iteration information.

Accessing ModelData Objects

Parameter and array objects can be access from the ModelData structure via the obj_dict field.

using EarthBox

eb = EarthBoxState(xnum=100, ynum=100, xsize=100000, ysize=100000.0, dx_marker=100.0, dy_marker=100.0);

model = eb.model_manager.model;
xnum = model.obj_dict["xnum"];
marker_x = model.obj_dict["marker_x"];
marker_y = model.obj_dict["marker_y"];

x = marker_x.array[50]
array_units = marker_x.units
array_description = marker_x.description

value = xnum.value
units = xnum.units
parameter_description = xnum.description

nested dot access

Model objects can also be accessed via nested dot structures. For example, the marker x-coordinates can be accessed via model.markers.arrays.location.marker_x.array. The nested dot access patterns are defined for each objects in a model data collection. See [MarkerContainer.ArrayCollection.Location][@ref ModelDataContainer.MarkerContainer.ArrayCollection.Location] for an example of nested dot access definitions.